db/requirements.txt
fastapi>=0.103.0
uvicorn[all]>=0.22.0
requests
loguru>=0.7.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
psycopg2>=2.9.6
pandas>=2.0.3
sqlalchemy>=2.0.29
pydantic>=2.0.0
aiohttp>=3.8.4
sentry-sdk
plotly==5.16.1
kaleido==0.2.1
goplus==0.2.1